fix(database): cast size_bytes as bigint in sqlc queries

+++ internal/storage/database/documents.sql.go
@@ -28,7 +28,7 @@ VALUES
     $2::text,
     $3::text,
     $4::text,
-    $5::int,
+    $5::bigint,
     $6::text,
     $7::text[]
   )
